+++ src/main/java/insane96mcp/iguanatweaksreborn/effect/Vigour.java
@@ -7,17 +7,17 @@ import net.minecraft.world.entity.player.Player;
 /**
  * Reduces exhaustion by 20% per level
  */
-public class WellRested extends ITMobEffect {
-    public WellRested(MobEffectCategory typeIn, int liquidColorIn) {
+public class Vigour extends ITMobEffect {
+    public Vigour(MobEffectCategory typeIn, int liquidColorIn) {
         super(typeIn, liquidColorIn);
     }
 
     public static float decreaseExhaustionConsumption(Player player, float exhaustion) {
-        if (!player.hasEffect(ITMobEffects.WELL_RESTED.get()))
+        if (!player.hasEffect(ITMobEffects.VIGOUR.get()))
             return exhaustion;
 
         //noinspection ConstantConditions
-        int amp = player.getEffect(ITMobEffects.WELL_RESTED.get()).getAmplifier() + 1;
-        return exhaustion * Math.max(0, 1 - amp * 0.2f);
+        int amp = player.getEffect(ITMobEffects.VIGOUR.get()).getAmplifier() + 1;
+        return exhaustion * Math.max(0, 1 - amp * 0.25f);
     }
 }