+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Upcoming
+* Renamed Well Rested to Vigour
+  * Vigour now reduces more hunger consumption (20% -> 25% per level)
+  * Can now be obtained with a Beacon when the Regeneration effect is active
+  * Can still be obtained by sleeping when not tired
+
 ## 2.13.4
 * Fixed well fed not applying
 * Fixed sleeping when Tired effect was present not setting respawn point