+++ changelog.md
@@ -5,6 +5,7 @@
   * Vigour now reduces more hunger consumption (~~20%~~ -> 25% per level)
   * Can now be obtained with a Beacon when the Regeneration effect is active
   * Can still be obtained by sleeping when not tired
+* Item modifiers now require UUIDs (Fixing multiple modifiers not stacking)
 
 ## 2.13.4
 * Fixed well fed not applying